refactor(projection): rename tier1 -> summarizer, Tier variants by function
