+++ crates/forge_app/src/orch_spec/orch_setup.rs
@@ -101,7 +101,7 @@ pub struct TestOutput {
     pub conversation_history: Vec<Conversation>,
     pub chat_responses: Vec<anyhow::Result<ChatResponse>>,
     /// Projected context the model saw per dispatch. Canonical-only
-    /// inspection can't distinguish pass-through from a tier-1 splice.
+    /// inspection can't distinguish pass-through from a summariser splice.
     pub outbound_contexts: Vec<forge_domain::Context>,
 }
