+++ crates/forge_app/src/orch_spec/orch_spec.rs
@@ -739,15 +739,15 @@ async fn test_projection_no_op_when_threshold_unset() {
     );
 }
 
-/// Guards the immutable-history invariant at the orch level: a tier-1
-/// projection that produces summary frames for the dispatch must not
-/// leak those frames into the persisted canonical.
+/// Guards the immutable-history invariant at the orch level: a
+/// summariser projection that produces summary frames for the dispatch
+/// must not leak those frames into the persisted canonical.
 #[tokio::test]
-async fn test_tier1_projection_does_not_mutate_canonical() {
+async fn test_summarizer_projection_does_not_mutate_canonical() {
     use forge_domain::{Agent, AgentId, Compact, ProviderId, Template};
     let mut compact = Compact::new();
-    // Any positive token count trips the token threshold so tier-1
-    // definitely fires on this tiny fixture.
+    // Any positive token count trips the token threshold so the
+    // summariser definitely fires on this tiny fixture.
     compact.token_threshold = Some(1);
     compact.message_threshold = Some(2);
     // Large cap keeps the slide step dormant — canonical leakage is
@@ -783,7 +783,7 @@ async fn test_tier1_projection_does_not_mutate_canonical() {
         .any(|content| content.contains("<summary"));
     assert!(
         !canonical_has_summary,
-        "canonical must not be mutated by tier-1 projection"
+        "canonical must not be mutated by summariser projection"
     );
 }
