+++ crates/forge_app/src/projection/summarizer.rs
